pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rungpt"
version = "0.1.0"
description = "一个强大的 AI Agent 框架，支持多种 Agent 类型、工具调用、记忆管理和上下文工程"
readme = "README.md"
authors = [
    {name = "RunGPT Team", email = "contact@rungpt.dev"}
]
license = {text = "MIT"}
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
keywords = ["ai", "agent", "llm", "chatbot", "react", "planner", "executor"]

dependencies = [
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/HemuCoder/rungpt"
Documentation = "https://github.com/HemuCoder/rungpt#readme"
Repository = "https://github.com/HemuCoder/rungpt"
"Bug Tracker" = "https://github.com/HemuCoder/rungpt/issues"

[tool.setuptools]
packages = ["rungpt", "rungpt.agent", "rungpt.context", "rungpt.models", "rungpt.threads", "rungpt.tools"]

[tool.setuptools.package-data]
rungpt = ["context/prompts/*.txt"]

[tool.black]
line-length = 100
target-version = ['py38']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

